+++ components/notification-center/App.tsx
@@ -3,19 +3,28 @@ import useNotifications from "../shared/hooks/useNotifications";
 import NotificationCenter from "./components/NotificationCenter";
 import NotificationsPanel from "./components/NotificationsPanel";
 import NotificationDetailPanel from "./components/NotificationDetailPanel";
-import { useState } from "react";
+import { usePubSub } from "../shared/hooks/finsemble-hooks";
+
+import { useState, useEffect } from "react";
 
 const App = (): React.ReactElement => {
 	const { notifications, doAction } = useNotifications();
 	const [activeNotification, setActiveNotification] = useState();
+	const pubSubTopic = "notification-ui";
+	const [notificationSubscribeMessage, notificationsPublish] = usePubSub(pubSubTopic);
+
+	useEffect(() => {
+		if ("showCenter" in notificationSubscribeMessage) {
+			notificationSubscribeMessage.showCenter ? finsembleWindow.show(null) : finsembleWindow.hide();
+		}
+	}, [notificationSubscribeMessage]);
 
 	return (
 		<div id="app">
 			<NotificationCenter title="Notification Center">
 				<div id="main-content">
 					{notifications.length === 0 ? (
-						// TODO:Add animated no notification component
-						<p>Good news no notifications!</p>
+						<p>You do not have any notifications!</p>
 					) : (
 						<>
 							<NotificationsPanel notifications={notifications} setActiveNotification={setActiveNotification} />