+++ components/shared/hooks/useNotifications.ts
@@ -1,21 +1,16 @@
-import { ToggleComponent } from "./../../../types/Notification-definitions/NotificationHookTypes.d";
-import { FSBL } from "./../../../types/FSBL-definitions/globals.d";
-/* eslint-disable @typescript-eslint/ban-ts-ignore */
-import * as react from "react";
-import { WindowIdentifier } from "../../../types/FSBL-definitions/globals";
+import * as React from "react";
 import INotification from "../../../types/Notification-definitions/INotification";
 import Subscription from "../../../types/Notification-definitions/Subscription";
 import NotificationClient from "../../../services/notification/notificationClient";
 import Filter from "../../../types/Notification-definitions/Filter";
-import { SpawnParams } from "../../../types/FSBL-definitions/services/window/Launcher/launcher";
 import WindowConfig, { NotificationsConfig } from "../../../types/Notification-definitions/NotificationConfig";
 import IFilter from "../../../types/Notification-definitions/IFilter";
 import { NotificationGroupList } from "../../../types/Notification-definitions/NotificationHookTypes";
-import _get = require("lodash/get");
+import _get from "lodash.get";
 
-const FSBL = window.FSBL;
+const { useReducer, useEffect } = React;
 
-const { LauncherClient, WindowClient } = FSBL.Clients;
+const { WindowClient } = FSBL.Clients;
 
 const initialState: any = { notifications: [] };
 
@@ -58,8 +53,8 @@ function reducer(state: { notifications: INotification[] }, action: { type: stri
 	}
 }
 
-export default function useNotifications() {
-	const [state, dispatch] = react.useReducer(reducer, initialState);
+export default function useNotifications(params: any = {}) {
+	const [state, dispatch] = useReducer(reducer, initialState);
 
 	let NOTIFICATION_CLIENT: NotificationClient = null;
 
@@ -79,7 +74,7 @@ export default function useNotifications() {
 	};
 
 	// start receiving Notifications and putting them in state
-	react.useEffect(() => {
+	useEffect(() => {
 		// eslint-disable-next-line @typescript-eslint/no-use-before-define
 		const subscribe = init();
 		return () => {
@@ -121,7 +116,6 @@ export default function useNotifications() {
 			arr
 				.map(
 					(notification: INotification): INotification["type"] =>
-						//TODO: this does not work as expected without the ignore
 						//@ts-ignore
 						notification[type]
 				)
@@ -146,8 +140,6 @@ export default function useNotifications() {
 		NOTIFICATION_CLIENT = new NotificationClient();
 		return NOTIFICATION_CLIENT.fetchHistory(since, filter);
 	};
-
-	//TODO: move out to a Finsemble hook and import here
 	/**
 	 * Get Notification's config from
 	 * @param componentType Finsemble component type e.g "Welcome-Component"
@@ -158,65 +150,11 @@ export default function useNotifications() {
 		return _get(config, "window.data.notifications", null);
 	};
 
-	//TODO: move out to a Finsemble hook and import here
-	/*
-	Finsemble Window manipulation
-*/
-
-	const setWindowPosition = async (windowId: WindowIdentifier, windowShowParams: SpawnParams): Promise<any> => {
-		const { windowDescriptor: windowPosition } = (await LauncherClient.showWindow(windowId, windowShowParams)).data;
-		return windowPosition;
-	};
-
-	/**
-	 * Set the position of the notification drawer based on params
-	 * @param param0
-	 */
-	const setNotificationDrawerPosition = async (windowShowParams: SpawnParams) => {
-		const windowId: WindowIdentifier = finsembleWindow.windowOptions.windowIdentifier;
-		await setWindowPosition(windowId, windowShowParams);
-	};
-
-	const minimizeWindow = () => {
-		finsembleWindow.hide();
-	};
+	const notificationIsActive = (notification: INotification) =>
+		!notification.isSnoozed && !notification.isRead && !notification.isDeleted;
 
 	const activeNotifications = (notifications: INotification[]) =>
-		notifications.filter(notification => !notification.isSnoozed && !notification.isRead);
-
-	const getWindowSpawnData = () => {
-		return WindowClient.getSpawnData();
-	};
-
-	const toggleComponent = async ({
-		windowName,
-		componentType,
-		showAction,
-		hideAction
-	}: ToggleComponent): Promise<void> => {
-		//set window.name to make the it easier to wrap (as name will be fixed)
-		const { wrap: component } = await FSBL.FinsembleWindow.wrap({ windowName }, (err: Error, wrap: object) =>
-			err ? console.error : wrap
-		);
-
-		//toggle visibility (onto monitor its being called on!)
-		component.isShowing({}, (err: Error, isShowing: boolean) => {
-			try {
-				if (isShowing) {
-					// default action
-					!hideAction ? component.hide() : hideAction();
-				} else {
-					// default action
-					!showAction
-						? LauncherClient.showWindow({ windowName, componentType }, { spawnIfNotFound: true })
-						: showAction();
-				}
-			} catch (error) {
-				console.error(error);
-				FSBL.Clients.Logger.error(error);
-			}
-		});
-	};
+		notifications.filter(notification => notificationIsActive(notification));
 
 	/**
 	 * Main init function to start the subscription
@@ -245,14 +183,21 @@ export default function useNotifications() {
 
 			subscription.filter = filter;
 
-			if (notificationConfig && notificationConfig.notificationsHistory) {
+			if (
+				(notificationConfig && notificationConfig.notificationsHistory) ||
+				_get(params, "config.notificationsHistory")
+			) {
 				// const { since, filter } = notificationConfig.notificationsHistory;
 				const pastNotifications = await getNotificationHistory();
 				addMultipleNotifications(pastNotifications);
 			}
 			subscription.onNotification = function(notification: INotification) {
 				// This function will be called when a notification arrives
-				addNotification(notification);
+				if (notification.isDeleted) {
+					removeNotification(notification);
+				} else {
+					addNotification(notification);
+				}
 			};
 
 			return NOTIFICATION_CLIENT.subscribe(
@@ -271,15 +216,12 @@ export default function useNotifications() {
 
 	return {
 		activeNotifications,
+		notificationIsActive,
 		doAction,
-		getWindowSpawnData,
 		getNotificationHistory,
 		groupNotificationsByType,
-		minimizeWindow,
 		notifications: state.notifications,
 		removeNotification,
-		setNotificationDrawerPosition,
-		getNotificationConfig,
-		toggleComponent
+		getNotificationConfig
 	};
 }