+++ types/FSBL-definitions/services/workspace/workspaceService.d.ts
@@ -1,7 +1,0 @@
-import { Workspace, ActiveWorkspace, GroupData } from "../../common/workspace";
-export declare type WorkspaceStateUpdate = {
-    activeWorkspace: ActiveWorkspace;
-    workspaces: Workspace[];
-    groups: Record<string, GroupData>;
-    reason: string;
-};