+++ types/Notification-definitions/NotificationConfig.d.ts
@@ -1,6 +1,4 @@
-import { SpawnParams } from "../FSBL-definitions/services/window/Launcher/launcher";
-
-// type SpawnParamsWithoutData = Omit<SpawnParams, "data">;
+import { SpawnParams } from "@chartiq/finsemble/dist/types/services/window/Launcher/launcher";
 
 export interface NotificationsConfig {
 	filter?: {
@@ -14,6 +12,11 @@ export interface NotificationsConfig {
 		animateOut: string;
 	};
 	notificationsHistory?: boolean;
+}
+
+export interface PurgeConfig {
+	maxNotificationsToRetain: number;
+	maxNotificationRetentionPeriodSeconds: number | false;
 }
 
 export default interface WindowConfig {