+++ types/Notification-definitions/SnoozeTimer.ts
@@ -4,5 +4,6 @@ import Timeout = NodeJS.Timeout;
 export default class SnoozeTimer implements ISnoozeTimer {
 	timeoutId: Timeout;
 	notificationId: string;
+	wakeEpochMilliseconds: number;
 	snoozeInterval: number;
 }